snapshot of progress
