+++ src/compile/nodes/Element.ts
@@ -113,7 +113,6 @@ export default class Element extends Node {
 	actions: Action[];
 	bindings: Binding[];
 	classes: Class[];
-	classDependencies: string[];
 	handlers: EventHandler[];
 	intro?: Transition;
 	outro?: Transition;
@@ -144,7 +143,6 @@ export default class Element extends Node {
 		this.actions = [];
 		this.bindings = [];
 		this.classes = [];
-		this.classDependencies = [];
 		this.handlers = [];
 
 		this.intro = null;
@@ -614,737 +612,6 @@ export default class Element extends Node {
 		}
 	}
 
-	init(
-		block: Block,
-		stripWhitespace: boolean,
-		nextSibling: Node
-	) {
-		if (this.name === 'slot' || this.name === 'option' || this.component.options.dev) {
-			this.cannotUseInnerHTML();
-		}
-
-		this.var = block.getUniqueName(
-			this.name.replace(/[^a-zA-Z0-9_$]/g, '_')
-		);
-
-		this.attributes.forEach(attr => {
-			if (
-				attr.chunks &&
-				attr.chunks.length &&
-				(attr.chunks.length > 1 || attr.chunks[0].type !== 'Text')
-			) {
-				this.parent.cannotUseInnerHTML();
-			}
-			if (attr.dependencies.size) {
-				block.addDependencies(attr.dependencies);
-
-				// special case — <option value={foo}> — see below
-				if (this.name === 'option' && attr.name === 'value') {
-					let select = this.parent;
-					while (select && (select.type !== 'Element' || select.name !== 'select')) select = select.parent;
-
-					if (select && select.selectBindingDependencies) {
-						select.selectBindingDependencies.forEach(prop => {
-							attr.dependencies.forEach((dependency: string) => {
-								this.component.indirectDependencies.get(prop).add(dependency);
-							});
-						});
-					}
-				}
-			}
-		});
-
-		this.actions.forEach(action => {
-			this.parent.cannotUseInnerHTML();
-			if (action.expression) {
-				block.addDependencies(action.expression.dependencies);
-			}
-		});
-
-		this.bindings.forEach(binding => {
-			this.parent.cannotUseInnerHTML();
-			block.addDependencies(binding.value.dependencies);
-		});
-
-		this.classes.forEach(classDir => {
-			this.parent.cannotUseInnerHTML();
-			if (classDir.expression) {
-				block.addDependencies(classDir.expression.dependencies);
-			}
-		});
-
-		this.handlers.forEach(handler => {
-			this.parent.cannotUseInnerHTML();
-			block.addDependencies(handler.dependencies);
-		});
-
-		if (this.intro || this.outro || this.animation || this.ref) {
-			this.parent.cannotUseInnerHTML();
-		}
-
-		if (this.intro) block.addIntro();
-		if (this.outro) block.addOutro();
-		if (this.animation) block.addAnimation();
-
-		const valueAttribute = this.attributes.find((attribute: Attribute) => attribute.name === 'value');
-
-		// special case — in a case like this...
-		//
-		//   <select bind:value='foo'>
-		//     <option value='{bar}'>bar</option>
-		//     <option value='{baz}'>baz</option>
-		//   </option>
-		//
-		// ...we need to know that `foo` depends on `bar` and `baz`,
-		// so that if `foo.qux` changes, we know that we need to
-		// mark `bar` and `baz` as dirty too
-		if (this.name === 'select') {
-			const binding = this.bindings.find(node => node.name === 'value');
-			if (binding) {
-				// TODO does this also apply to e.g. `<input type='checkbox' bind:group='foo'>`?
-				const dependencies = binding.value.dependencies;
-				this.selectBindingDependencies = dependencies;
-				dependencies.forEach((prop: string) => {
-					this.component.indirectDependencies.set(prop, new Set());
-				});
-			} else {
-				this.selectBindingDependencies = null;
-			}
-		}
-
-		const slot = this.getStaticAttributeValue('slot');
-		if (slot && this.hasAncestor('InlineComponent')) {
-			this.cannotUseInnerHTML();
-			this.slotted = true;
-			// TODO validate slots — no nesting, no dynamic names...
-			const component = this.findNearest(/^InlineComponent/);
-			component._slots.add(slot);
-		}
-
-		if (this.children.length) {
-			if (this.name === 'pre' || this.name === 'textarea') stripWhitespace = false;
-			this.initChildren(block, stripWhitespace, nextSibling);
-		}
-	}
-
-	build(
-		block: Block,
-		parentNode: string,
-		parentNodes: string
-	) {
-		const { component } = this;
-
-		if (this.name === 'slot') {
-			const slotName = this.getStaticAttributeValue('name') || 'default';
-			this.component.slots.add(slotName);
-		}
-
-		if (this.name === 'noscript') return;
-
-		const node = this.var;
-		const nodes = parentNodes && block.getUniqueName(`${this.var}_nodes`) // if we're in unclaimable territory, i.e. <head>, parentNodes is null
-
-		const slot = this.attributes.find((attribute: Node) => attribute.name === 'slot');
-		const prop = slot && quotePropIfNecessary(slot.chunks[0].data);
-		const initialMountNode = this.slotted ?
-			`${this.findNearest(/^InlineComponent/).var}._slotted${prop}` : // TODO this looks bonkers
-			parentNode;
-
-		block.addVariable(node);
-		const renderStatement = getRenderStatement(this.namespace, this.name);
-		block.builders.create.addLine(
-			`${node} = ${renderStatement};`
-		);
-
-		if (this.component.options.hydratable) {
-			if (parentNodes) {
-				block.builders.claim.addBlock(deindent`
-					${node} = ${getClaimStatement(component, this.namespace, parentNodes, this)};
-					var ${nodes} = @children(${this.name === 'template' ? `${node}.content` : node});
-				`);
-			} else {
-				block.builders.claim.addLine(
-					`${node} = ${renderStatement};`
-				);
-			}
-		}
-
-		if (initialMountNode) {
-			block.builders.mount.addLine(
-				`@append(${initialMountNode}, ${node});`
-			);
-
-			if (initialMountNode === 'document.head') {
-				block.builders.destroy.addLine(`@detachNode(${node});`);
-			}
-		} else {
-			block.builders.mount.addLine(`@insert(#target, ${node}, anchor);`);
-
-			// TODO we eventually need to consider what happens to elements
-			// that belong to the same outgroup as an outroing element...
-			block.builders.destroy.addConditional('detach', `@detachNode(${node});`);
-		}
-
-		// insert static children with textContent or innerHTML
-		if (!this.namespace && this.canUseInnerHTML && this.children.length > 0) {
-			if (this.children.length === 1 && this.children[0].type === 'Text') {
-				block.builders.create.addLine(
-					`${node}.textContent = ${stringify(this.children[0].data)};`
-				);
-			} else {
-				block.builders.create.addLine(
-					`${node}.innerHTML = ${stringify(this.children.map(toHTML).join(''))};`
-				);
-			}
-		} else {
-			this.children.forEach((child: Node) => {
-				child.build(block, this.name === 'template' ? `${node}.content` : node, nodes);
-			});
-		}
-
-		let hasHoistedEventHandlerOrBinding = (
-			//(this.hasAncestor('EachBlock') && this.bindings.length > 0) ||
-			this.handlers.some(handler => handler.shouldHoist)
-		);
-		const eventHandlerOrBindingUsesComponent = (
-			this.bindings.length > 0 ||
-			this.handlers.some(handler => handler.usesComponent)
-		);
-
-		const eventHandlerOrBindingUsesContext = (
-			this.bindings.some(binding => binding.usesContext) ||
-			this.handlers.some(handler => handler.usesContext)
-		);
-
-		if (hasHoistedEventHandlerOrBinding) {
-			const initialProps: string[] = [];
-			const updates: string[] = [];
-
-			if (eventHandlerOrBindingUsesComponent) {
-				const component = block.alias('component');
-				initialProps.push(component === 'component' ? 'component' : `component: ${component}`);
-			}
-
-			if (eventHandlerOrBindingUsesContext) {
-				initialProps.push(`ctx`);
-				block.builders.update.addLine(`${node}._svelte.ctx = ctx;`);
-				block.maintainContext = true;
-			}
-
-			if (initialProps.length) {
-				block.builders.hydrate.addBlock(deindent`
-					${node}._svelte = { ${initialProps.join(', ')} };
-				`);
-			}
-		} else {
-			if (eventHandlerOrBindingUsesContext) {
-				block.maintainContext = true;
-			}
-		}
-
-		this.addBindings(block);
-		this.addEventHandlers(block);
-		if (this.ref) this.addRef(block);
-		this.addAttributes(block);
-		this.addTransitions(block);
-		this.addAnimation(block);
-		this.addActions(block);
-		this.addClasses(block);
-
-		if (this.initialUpdate) {
-			block.builders.mount.addBlock(this.initialUpdate);
-		}
-
-		if (nodes) {
-			block.builders.claim.addLine(
-				`${nodes}.forEach(@detachNode);`
-			);
-		}
-
-		function toHTML(node: Element | Text) {
-			if (node.type === 'Text') {
-				return node.parent &&
-					node.parent.type === 'Element' &&
-					(node.parent.name === 'script' || node.parent.name === 'style')
-					? node.data
-					: escapeHTML(node.data);
-			}
-
-			if (node.name === 'noscript') return '';
-
-			let open = `<${node.name}`;
-
-			node.attributes.forEach((attr: Node) => {
-				open += ` ${fixAttributeCasing(attr.name)}${stringifyAttributeValue(attr.chunks)}`
-			});
-
-			if (isVoidElementName(node.name)) return open + '>';
-
-			return `${open}>${node.children.map(toHTML).join('')}</${node.name}>`;
-		}
-
-		if (this.component.options.dev) {
-			const loc = this.component.locate(this.start);
-			block.builders.hydrate.addLine(
-				`@addLoc(${this.var}, ${this.component.fileVar}, ${loc.line}, ${loc.column}, ${this.start});`
-			);
-		}
-	}
-
-	addBindings(
-		block: Block
-	) {
-		if (this.bindings.length === 0) return;
-
-		if (this.name === 'select' || this.isMediaNode()) this.component.target.hasComplexBindings = true;
-
-		const needsLock = this.name !== 'input' || !/radio|checkbox|range|color/.test(this.getStaticAttributeValue('type'));
-
-		// TODO munge in constructor
-		const mungedBindings = this.bindings.map(binding => binding.munge(block));
-
-		const lock = mungedBindings.some(binding => binding.needsLock) ?
-			block.getUniqueName(`${this.var}_updating`) :
-			null;
-
-		if (lock) block.addVariable(lock, 'false');
-
-		const groups = events
-			.map(event => {
-				return {
-					events: event.eventNames,
-					bindings: mungedBindings.filter(binding => event.filter(this, binding.name))
-				};
-			})
-			.filter(group => group.bindings.length);
-
-		groups.forEach(group => {
-			const handler = block.getUniqueName(`${this.var}_${group.events.join('_')}_handler`);
-
-			const needsLock = group.bindings.some(binding => binding.needsLock);
-
-			group.bindings.forEach(binding => {
-				if (!binding.updateDom) return;
-
-				const updateConditions = needsLock ? [`!${lock}`] : [];
-				if (binding.updateCondition) updateConditions.push(binding.updateCondition);
-
-				block.builders.update.addLine(
-					updateConditions.length ? `if (${updateConditions.join(' && ')}) ${binding.updateDom}` : binding.updateDom
-				);
-			});
-
-			const usesContext = group.bindings.some(binding => binding.handler.usesContext);
-			const usesState = group.bindings.some(binding => binding.handler.usesState);
-			const usesStore = group.bindings.some(binding => binding.handler.usesStore);
-			const mutations = group.bindings.map(binding => binding.handler.mutation).filter(Boolean).join('\n');
-
-			const props = new Set();
-			const storeProps = new Set();
-			group.bindings.forEach(binding => {
-				binding.handler.props.forEach(prop => {
-					props.add(prop);
-				});
-
-				binding.handler.storeProps.forEach(prop => {
-					storeProps.add(prop);
-				});
-			}); // TODO use stringifyProps here, once indenting is fixed
-
-			// media bindings — awkward special case. The native timeupdate events
-			// fire too infrequently, so we need to take matters into our
-			// own hands
-			let animation_frame;
-			if (group.events[0] === 'timeupdate') {
-				animation_frame = block.getUniqueName(`${this.var}_animationframe`);
-				block.addVariable(animation_frame);
-			}
-
-			block.builders.init.addBlock(deindent`
-				function ${handler}() {
-					${
-						animation_frame && deindent`
-							cancelAnimationFrame(${animation_frame});
-							if (!${this.var}.paused) ${animation_frame} = requestAnimationFrame(${handler});`
-					}
-					${usesStore && `var $ = #component.store.get();`}
-					${needsLock && `${lock} = true;`}
-					${mutations.length > 0 && mutations}
-					${props.size > 0 && `#component.set({ ${Array.from(props).join(', ')} });`}
-					${storeProps.size > 0 && `#component.store.set({ ${Array.from(storeProps).join(', ')} });`}
-					${needsLock && `${lock} = false;`}
-				}
-			`);
-
-			group.events.forEach(name => {
-				if (name === 'resize') {
-					// special case
-					const resize_listener = block.getUniqueName(`${this.var}_resize_listener`);
-					block.addVariable(resize_listener);
-
-					block.builders.mount.addLine(
-						`${resize_listener} = @addResizeListener(${this.var}, ${handler});`
-					);
-
-					block.builders.destroy.addLine(
-						`${resize_listener}.cancel();`
-					);
-				} else {
-					block.builders.hydrate.addLine(
-						`@addListener(${this.var}, "${name}", ${handler});`
-					);
-
-					block.builders.destroy.addLine(
-						`@removeListener(${this.var}, "${name}", ${handler});`
-					);
-				}
-			});
-
-			const allInitialStateIsDefined = group.bindings
-				.map(binding => `'${binding.object}' in ctx`)
-				.join(' && ');
-
-			if (this.name === 'select' || group.bindings.find(binding => binding.name === 'indeterminate' || binding.isReadOnlyMediaAttribute)) {
-				this.component.target.hasComplexBindings = true;
-
-				block.builders.hydrate.addLine(
-					`if (!(${allInitialStateIsDefined})) #component.root._beforecreate.push(${handler});`
-				);
-			}
-
-			if (group.events[0] === 'resize') {
-				this.component.target.hasComplexBindings = true;
-
-				block.builders.hydrate.addLine(
-					`#component.root._beforecreate.push(${handler});`
-				);
-			}
-		});
-
-		this.initialUpdate = mungedBindings.map(binding => binding.initialUpdate).filter(Boolean).join('\n');
-	}
-
-	addAttributes(block: Block) {
-		if (this.attributes.find(attr => attr.type === 'Spread')) {
-			this.addSpreadAttributes(block);
-			return;
-		}
-
-		this.attributes.forEach((attribute: Attribute) => {
-			if (attribute.name === 'class' && attribute.isDynamic) {
-				this.classDependencies.push(...attribute.dependencies);
-			}
-			attribute.render(block);
-		});
-	}
-
-	addSpreadAttributes(block: Block) {
-		const levels = block.getUniqueName(`${this.var}_levels`);
-		const data = block.getUniqueName(`${this.var}_data`);
-
-		const initialProps = [];
-		const updates = [];
-
-		this.attributes
-			.filter(attr => attr.type === 'Attribute' || attr.type === 'Spread')
-			.forEach(attr => {
-				const condition = attr.dependencies.size > 0
-					? `(${[...attr.dependencies].map(d => `changed.${d}`).join(' || ')})`
-					: null;
-
-				if (attr.isSpread) {
-					const { snippet, dependencies } = attr.expression;
-
-					initialProps.push(snippet);
-
-					updates.push(condition ? `${condition} && ${snippet}` : snippet);
-				} else {
-					const snippet = `{ ${quoteNameIfNecessary(attr.name)}: ${attr.getValue()} }`;
-					initialProps.push(snippet);
-
-					updates.push(condition ? `${condition} && ${snippet}` : snippet);
-				}
-			});
-
-		block.builders.init.addBlock(deindent`
-			var ${levels} = [
-				${initialProps.join(',\n')}
-			];
-
-			var ${data} = {};
-			for (var #i = 0; #i < ${levels}.length; #i += 1) {
-				${data} = @assign(${data}, ${levels}[#i]);
-			}
-		`);
-
-		block.builders.hydrate.addLine(
-			`@setAttributes(${this.var}, ${data});`
-		);
-
-		block.builders.update.addBlock(deindent`
-			@setAttributes(${this.var}, @getSpreadUpdate(${levels}, [
-				${updates.join(',\n')}
-			]));
-		`);
-	}
-
-	addEventHandlers(block: Block) {
-		const { component } = this;
-
-		this.handlers.forEach(handler => {
-			const isCustomEvent = component.events.has(handler.name);
-
-			if (handler.callee) {
-				handler.render(this.component, block, handler.shouldHoist);
-			}
-
-			const target = handler.shouldHoist ? 'this' : this.var;
-
-			// get a name for the event handler that is globally unique
-			// if hoisted, locally unique otherwise
-			const handlerName = (handler.shouldHoist ? component : block).getUniqueName(
-				`${handler.name.replace(/[^a-zA-Z0-9_$]/g, '_')}_handler`
-			);
-
-			const component_name = block.alias('component'); // can't use #component, might be hoisted
-
-			// create the handler body
-			const handlerBody = deindent`
-				${handler.shouldHoist && (
-					handler.usesComponent || handler.usesContext
-						? `const { ${[handler.usesComponent && 'component', handler.usesContext && 'ctx'].filter(Boolean).join(', ')} } = ${target}._svelte;`
-						: null
-				)}
-
-				${handler.snippet ?
-					handler.snippet :
-					`${component_name}.fire("${handler.name}", event);`}
-			`;
-
-			if (isCustomEvent) {
-				block.addVariable(handlerName);
-
-				block.builders.hydrate.addBlock(deindent`
-					${handlerName} = %events-${handler.name}.call(${component_name}, ${this.var}, function(event) {
-						${handlerBody}
-					});
-				`);
-
-				block.builders.destroy.addLine(deindent`
-					${handlerName}.destroy();
-				`);
-			} else {
-				const handlerFunction = deindent`
-					function ${handlerName}(event) {
-						${handlerBody}
-					}
-				`;
-
-				if (handler.shouldHoist) {
-					component.target.blocks.push(handlerFunction);
-				} else {
-					block.builders.init.addBlock(handlerFunction);
-				}
-
-				block.builders.hydrate.addLine(
-					`@addListener(${this.var}, "${handler.name}", ${handlerName});`
-				);
-
-				block.builders.destroy.addLine(
-					`@removeListener(${this.var}, "${handler.name}", ${handlerName});`
-				);
-			}
-		});
-	}
-
-	addRef(block: Block) {
-		const ref = `#component.refs.${this.ref.name}`;
-
-		block.builders.mount.addLine(
-			`${ref} = ${this.var};`
-		);
-
-		block.builders.destroy.addLine(
-			`if (${ref} === ${this.var}) ${ref} = null;`
-		);
-	}
-
-	addTransitions(
-		block: Block
-	) {
-		const { intro, outro } = this;
-
-		if (!intro && !outro) return;
-
-		if (intro === outro) {
-			const name = block.getUniqueName(`${this.var}_transition`);
-			const snippet = intro.expression
-				? intro.expression.snippet
-				: '{}';
-
-			block.addVariable(name);
-
-			const fn = `%transitions-${intro.name}`;
-
-			block.builders.intro.addConditional(`#component.root._intro`, deindent`
-				if (${name}) ${name}.invalidate();
-
-				#component.root._aftercreate.push(() => {
-					if (!${name}) ${name} = @wrapTransition(#component, ${this.var}, ${fn}, ${snippet}, true);
-					${name}.run(1);
-				});
-			`);
-
-			block.builders.outro.addBlock(deindent`
-				if (!${name}) ${name} = @wrapTransition(#component, ${this.var}, ${fn}, ${snippet}, false);
-				${name}.run(0, () => {
-					#outrocallback();
-					${name} = null;
-				});
-			`);
-
-			block.builders.destroy.addConditional('detach', `if (${name}) ${name}.abort();`);
-		} else {
-			const introName = intro && block.getUniqueName(`${this.var}_intro`);
-			const outroName = outro && block.getUniqueName(`${this.var}_outro`);
-
-			if (intro) {
-				block.addVariable(introName);
-				const snippet = intro.expression
-					? intro.expression.snippet
-					: '{}';
-
-				const fn = `%transitions-${intro.name}`; // TODO add built-in transitions?
-
-				if (outro) {
-					block.builders.intro.addBlock(deindent`
-						if (${introName}) ${introName}.abort(1);
-						if (${outroName}) ${outroName}.abort(1);
-					`);
-				}
-
-				block.builders.intro.addConditional(`#component.root._intro`, deindent`
-					#component.root._aftercreate.push(() => {
-						${introName} = @wrapTransition(#component, ${this.var}, ${fn}, ${snippet}, true);
-						${introName}.run(1);
-					});
-				`);
-			}
-
-			if (outro) {
-				block.addVariable(outroName);
-				const snippet = outro.expression
-					? outro.expression.snippet
-					: '{}';
-
-				const fn = `%transitions-${outro.name}`;
-
-				block.builders.intro.addBlock(deindent`
-					if (${outroName}) ${outroName}.abort(1);
-				`);
-
-				// TODO hide elements that have outro'd (unless they belong to a still-outroing
-				// group) prior to their removal from the DOM
-				block.builders.outro.addBlock(deindent`
-					${outroName} = @wrapTransition(#component, ${this.var}, ${fn}, ${snippet}, false);
-					${outroName}.run(0, #outrocallback);
-				`);
-
-				block.builders.destroy.addConditional('detach', `if (${outroName}) ${outroName}.abort();`);
-			}
-		}
-	}
-
-	addAnimation(block: Block) {
-		if (!this.animation) return;
-
-		const rect = block.getUniqueName('rect');
-		const animation = block.getUniqueName('animation');
-
-		block.addVariable(rect);
-		block.addVariable(animation);
-
-		block.builders.measure.addBlock(deindent`
-			${rect} = ${this.var}.getBoundingClientRect();
-		`);
-
-		block.builders.fix.addBlock(deindent`
-			@fixPosition(${this.var});
-			if (${animation}) ${animation}.stop();
-		`);
-
-		const params = this.animation.expression ? this.animation.expression.snippet : '{}';
-		block.builders.animate.addBlock(deindent`
-			if (${animation}) ${animation}.stop();
-			${animation} = @wrapAnimation(${this.var}, ${rect}, %animations-${this.animation.name}, ${params});
-		`);
-	}
-
-	addActions(block: Block) {
-		this.actions.forEach(action => {
-			const { expression } = action;
-			let snippet, dependencies;
-			if (expression) {
-				snippet = expression.snippet;
-				dependencies = expression.dependencies;
-			}
-
-			const name = block.getUniqueName(
-				`${action.name.replace(/[^a-zA-Z0-9_$]/g, '_')}_action`
-			);
-
-			block.addVariable(name);
-			const fn = `%actions-${action.name}`;
-
-			block.builders.mount.addLine(
-				`${name} = ${fn}.call(#component, ${this.var}${snippet ? `, ${snippet}` : ''}) || {};`
-			);
-
-			if (dependencies && dependencies.size > 0) {
-				let conditional = `typeof ${name}.update === 'function' && `;
-				const deps = [...dependencies].map(dependency => `changed.${dependency}`).join(' || ');
-				conditional += dependencies.size > 1 ? `(${deps})` : deps;
-
-				block.builders.update.addConditional(
-					conditional,
-					`${name}.update.call(#component, ${snippet});`
-				);
-			}
-
-			block.builders.destroy.addLine(
-				`if (typeof ${name}.destroy === 'function') ${name}.destroy.call(#component);`
-			);
-		});
-	}
-
-	addClasses(block: Block) {
-		this.classes.forEach(classDir => {
-			const { expression, name } = classDir;
-			let snippet, dependencies;
-			if (expression) {
-				snippet = expression.snippet;
-				dependencies = expression.dependencies;
-			} else {
-				snippet = `ctx${quotePropIfNecessary(name)}`;
-				dependencies = [name];
-			}
-			const updater = `@toggleClass(${this.var}, "${name}", ${snippet});`;
-
-			block.builders.hydrate.addLine(updater);
-
-			if ((dependencies && dependencies.size > 0) || this.classDependencies.length) {
-				const allDeps = this.classDependencies.concat(...dependencies);
-				const deps = allDeps.map(dependency => `changed${quotePropIfNecessary(dependency)}`).join(' || ');
-				const condition = allDeps.length > 1 ? `(${deps})` : deps;
-
-				block.builders.update.addConditional(
-					condition,
-					updater
-				);
-			}
-		});
-	}
-
 	getStaticAttributeValue(name: string) {
 		const attribute = this.attributes.find(
 			(attr: Attribute) => attr.type === 'Attribute' && attr.name.toLowerCase() === name
@@ -1414,32 +681,6 @@ function getRenderStatement(
 	}
 
 	return `@createElement("${name}")`;
-}
-
-function getClaimStatement(
-	component: Component,
-	namespace: string,
-	nodes: string,
-	node: Node
-) {
-	const attributes = node.attributes
-		.filter((attr: Node) => attr.type === 'Attribute')
-		.map((attr: Node) => `${quoteNameIfNecessary(attr.name)}: true`)
-		.join(', ');
-
-	const name = namespace ? node.name : node.name.toUpperCase();
-
-	return `@claimElement(${nodes}, "${name}", ${attributes
-		? `{ ${attributes} }`
-		: `{}`}, ${namespace === namespaces.svg ? true : false})`;
-}
-
-function stringifyAttributeValue(value: Node[] | true) {
-	if (value === true) return '';
-	if (value.length === 0) return `=""`;
-
-	const data = value[0].data;
-	return `=${JSON.stringify(data)}`;
 }
 
 const events = [
