+++ src/compile/render-dom/wrappers/MustacheTag.ts
@@ -1,0 +1,29 @@
+import Renderer from '../Renderer';
+import Block from '../Block';
+import Node from '../../nodes/shared/Node';
+import Tag from './shared/Tag';
+
+export default class MustacheTagWrapper extends Tag {
+	constructor(renderer: Renderer, block: Block, parent: Wrapper, node: Node) {
+		super(renderer, block, parent, node);
+		this.cannotUseInnerHTML();
+	}
+
+	render(block: Block, parentNode: string, parentNodes: string) {
+		const { init } = this.renameThisMethod(
+			block,
+			value => `@setData(${this.var}, ${value});`
+		);
+
+		block.addElement(
+			this.var,
+			`@createText(${init})`,
+			parentNodes && `@claimText(${parentNodes}, ${init})`,
+			parentNode
+		);
+	}
+
+	remount(name: string) {
+		return `@append(${name}._slotted.default, ${this.var});`;
+	}
+}
